First Commit on Chama Investments
A Chama is an informal cooperative society that is normally used to pool and invest savings by people in East Africa, and particularly Kenya.

Of late Chama's relied on the personal trust vested in the member to keep the funds. But as history has shown, that trust is usually broken. My aim is to try and solve the problems inherent in the Chama societies using Radix.